pyproject.toml
[project]
name = "veo-prompts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-settings>=2.12.0",
    "anthropic>=0.40.0",
    "python-docx>=1.1.0",
    "playwright>=1.41.0",
    "playwright-stealth>=2.0.1",
]
